backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
pydantic==2.5.0
websockets==12.0
python-dotenv==1.0.0
aiohttp==3.9.1
pyyaml==6.0.1
